requirements.txt
PYMYSQL==1.1.0
pyyaml==6.0.1
django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
cryptography==42.0.0
validate-email==1.3
requests==2.28.2